requirements.txt
GitPython==3.1.32
requests==2.31.0
toml==0.10.2
